refactor(mf): separate entrypoint into local and remote
