+++ src/App.tsx
@@ -1,26 +1,9 @@
 import React from 'react';
-import { RouterProvider } from 'react-router-dom';
 import { createRoot } from 'react-dom/client';
-import { QueryClient, QueryClientProvider } from 'react-query';
-import { ReactQueryDevtools } from 'react-query/devtools';
 
-import router from './routes';
-
-const queryClient = new QueryClient();
-
-const Root = () => (
-  <React.StrictMode>
-    <QueryClientProvider client={queryClient}>
-      <RouterProvider router={router} />
-      <ReactQueryDevtools initialIsOpen={false} />
-    </QueryClientProvider>
-  </React.StrictMode>
-);
+import Root from './root';
 
 // local development
 const container = document.getElementById('app');
 const root = createRoot(container!);
 root.render(<Root />);
-
-// online
-// export default Root;
